working GraphQL server
