--- src/index.js
@@ -1,15 +1,0 @@
-const { ApolloServer } = require("apollo-server");
-const typeDefs = require("./schema");
-const resolvers = require("./resolvers");
-const models = require("../models");
-const { PORT } = require("../config/constants");
-
-const server = new ApolloServer({
-  typeDefs,
-  resolvers,
-  context: { models },
-});
-
-server
-  .listen()
-  .then(({ url }) => console.log(`Listening on port: ${PORT}`, models));
